+++ docs/pages/_app.tsx
@@ -1,8 +1,23 @@
 import type { AppProps } from 'next/app'
-import '@artemy-tech/datepicker/styles'
+import { useRouter } from 'next/router'
+import { ThemeProvider } from 'next-themes'
+import '@artemy-tech/rtdp/styles'
 import '../styles/global.css'
 import '../styles/nextra-overrides.css'
 
 export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  const router = useRouter()
+  const isDocs = router.pathname.startsWith('/docs')
+
+  if (isDocs) return <Component {...pageProps} />
+
+  return (
+    <ThemeProvider
+      attribute="class"
+      defaultTheme="dark"
+      disableTransitionOnChange
+    >
+      <Component {...pageProps} />
+    </ThemeProvider>
+  )
 }